requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aniso8601==9.0.1
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
flask-restx==1.1.0
frozenlist==1.3.3
future==0.18.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==2.4.2
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.7
OpenCC==1.1.6
psutil==5.9.5
pytz==2023.3
requests==2.28.2
six==1.16.0
tqdm==4.65.0
urllib3==1.26.16
waitress==2.1.2
Werkzeug==2.3.4
yarl==1.9.2
psycopg2-binary==2.9.9
cryptography==41.0.5
